Tested program trice
In the first test, there was an infinite loop from the view function.
After making changes to it, i ran the second test.

The second test had an infinite loop from the dequeue function

I reversed many of the changes i recently added to the program and executed the 3rd (+ a few more) tests.

They ran successfully